Point system redesign, Discord bot overhaul, debug logging toggle
Point system redesign:
- Points flow: earned → WeekEvents only (no duplication to month/year)
- End of week: auto-detected by ISO week change, cascade Weekly → Monthly
  via MergePoints + MergeScores, issue rewards if WeeklyRewardsEnabled
- End of month: cascade Monthly → Yearly, issue rewards if MonthlyRewardsEnabled
- End of year: clear Yearly, issue rewards if YearlyRewardsEnabled
- NexusManager: AddPointEvent weekly-only, date-based bucket placement
  for incoming Nexus points (IsCurrentWeek/Month/Year), cross-bucket dedup
- All score lists: int → ulong for overflow protection
- ResetEvent rewritten: ShouldProcessWeekly/Monthly/Yearly auto-detection
  via tracking fields, ForceSave after each step
- Config: removed WeeklyResetDay, MonthlyResetDay, YearlyResetMonth/Day,
  renamed enable flags to *RewardsEnabled, added 5 tracking fields
- Scores tab UI: removed date controls, renamed checkboxes to reward toggles
- ZonePointEvent: dead suits (character.IsDead) excluded from capture/decay.
  Piloted grids required for capture/decay (GetFatBlocks<MyCockpit>().Any).
  Auto-decay persists faction identity through empty zone.

Discord bot overhaul:
- Live scoreboard renamed "Points So Far", channel "points-this-week".
  Removed timestamp, embed-only (no plain content), title omitted.
- Zone channels: per-zone embeds posted with history (no editing), 10s rate
  limit per channel. Shows capture state, faction tag+name, points, time
  remaining. Skips Neutral (only posts on faction activity). Eviction
  warnings use Discord <t:timestamp:R> format, Warning10 removed (duplicate).
- Channel existence checks before every send for rewards, scoreboard, zones.
  Deleted channels are recreated with correct name + permissions.
- Message recovery: if scoreboard edit fails (message deleted), channel is
  deleted and recreated, new message ID saved.
- @everyone deny overwrites applied on channel creation and verified on
  every bot connect (VIEW_CHANNEL + READ_MESSAGE_HISTORY allowed, all else
  denied for @everyone). Admins inherit full access via server role.
- Configurable update interval (DiscordUpdateIntervalSeconds, default 30)
  and channel prefix (DiscordChannelPrefix, default "zone-"). UI controls
  in Discord tab.
- DiscordRestClient: added GetChannelAsync, ModifyChannelAsync,
  ModifyChannelPermissionsAsync, DeleteChannelAsync.
- DiscordModels: added PermissionOverwrite class, channel overwrites.

Debug logging toggle:
- KoTHLog static wrapper: Info/Warn gated on DebugLoggingEnabled config flag,
  Error always passes through
- DebugLoggingEnabled (bool, default false) config setting + UI checkbox
- All 77 Log.Info/Log.Warn calls migrated to KoTHLog.Info(Log, ...) /
  KoTHLog.Warn(Log, ...) across 11 files
- Reward commands use Log.Info directly (not KoTHLog) — always logged

Reward command execution fix:
- RunTorchCommand now calls CommandManager.HandleCommandFromServer,
  executing Torch commands with full admin privileges
- No automatic ! prefix added (admins supply correct server prefix)

QuestManager fixes:
- Grid power check: grid.GridSystems.ResourceDistributor
  .ResourceStateByType(ElectricityId) != NoPower — excludes debris
- Preview/projection skip: grid.Physics == null
- Non-character/non-grid entities excluded with else continue gate

+++ SenX KOTH Plugin/SenX KOTH Plugin/Utils/ScoreManager.cs
@@ -38,9 +38,9 @@ namespace SenX_KOTH_Plugin.Utils
 
     public sealed class ScoreFile : ViewModel
     {
-        public List<KeyValuePair<string, int>> WeekScores { get => field; set => SetValue(ref field, value); } = new();
-        public List<KeyValuePair<string, int>> MonthScores { get => field; set => SetValue(ref field, value); } = new();
-        public List<KeyValuePair<string, int>> YearScores { get => field; set => SetValue(ref field, value); } = new();
+        public List<KeyValuePair<string, ulong>> WeekScores { get => field; set => SetValue(ref field, value); } = new();
+        public List<KeyValuePair<string, ulong>> MonthScores { get => field; set => SetValue(ref field, value); } = new();
+        public List<KeyValuePair<string, ulong>> YearScores { get => field; set => SetValue(ref field, value); } = new();
     }
 }
